Rename integration tests to component tests
