+++ src/Shark.Fido2.sln
@@ -27,7 +27,7 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Shark.Fido2.Metadata.Core",
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Shark.Fido2.SqlServer", "Shark.Fido2.SqlServer\Shark.Fido2.SqlServer.csproj", "{D540F937-D674-408B-B5BE-54C5BAED77EF}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Shark.Fido2.Core.Integration.Tests", "..\tests\Shark.Fido2.Core.Integration.Tests\Shark.Fido2.Core.Integration.Tests.csproj", "{235ECDC5-C002-475E-BAEF-9590948C7CCB}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Shark.Fido2.Core.Component.Tests", "..\tests\Shark.Fido2.Core.Component.Tests\Shark.Fido2.Core.Component.Tests.csproj", "{235ECDC5-C002-475E-BAEF-9590948C7CCB}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Samples", "Samples", "{02EA681E-C7D8-13C7-8484-4AC65E1B71E8}"
 EndProject
