+++ tests/Shark.Fido2.Core.Component.Tests/Data/FidoU2fCreationOptions.json
