+++ tests/Shark.Fido2.Core.Component.Tests/Data/NoneAssertion.json
