+++ tests/Shark.Fido2.Core.Component.Tests/Data/PackedRequestOptions.json
