+++ tests/Shark.Fido2.Core.Component.Tests/PackedAttestationIntegrationTests.cs
@@ -7,18 +7,18 @@ using Shark.Fido2.Domain.Options;
 using Shark.Fido2.InMemory;
 using Shark.Fido2.Tests.Common.DataReaders;
 
-namespace Shark.Fido2.Core.Integration.Tests;
+namespace Shark.Fido2.Core.Component.Tests;
 
 /// <summary>
-/// Source: iOS 26.0 iPhone 14 authenticator.
+/// Source: Windows 10 Windows Hello authenticator.
 /// </summary>
 [TestFixture]
-internal class NoneAttestationIntegrationTests
+internal class PackedAttestationIntegrationTests
 {
-    private const string NoneAttestation = "NoneAttestation.json";
-    private const string NoneCreationOptions = "NoneCreationOptions.json";
-    private const string NoneAssertion = "NoneAssertion.json";
-    private const string NoneRequestOptions = "NoneRequestOptions.json";
+    private const string PackedAttestation = "PackedAttestation.json";
+    private const string PackedCreationOptions = "PackedCreationOptions.json";
+    private const string PackedAssertion = "PackedAssertion.json";
+    private const string PackedRequestOptions = "PackedRequestOptions.json";
 
     private ServiceProvider _serviceProvider = null!;
 
@@ -44,7 +44,7 @@ internal class NoneAttestationIntegrationTests
     }
 
     [Test]
-    public async Task BeginRegistration_WhenNoneAttestation_ThenReturnsSuccess()
+    public async Task BeginRegistration_WhenPackedAttestation_ThenReturnsSuccess()
     {
         // Arrange
         var attestation = _serviceProvider.GetRequiredService<IAttestation>();
@@ -85,13 +85,13 @@ internal class NoneAttestationIntegrationTests
     }
 
     [Test]
-    public async Task CompleteRegistration_WhenNoneAttestation_ThenReturnsSuccess()
+    public async Task CompleteRegistration_WhenPackedAttestation_ThenReturnsSuccess()
     {
         // Arrange
         var attestation = _serviceProvider.GetRequiredService<IAttestation>();
 
-        var attestationData = DataReader.ReadAttestationData(NoneAttestation);
-        var creationOptions = DataReader.ReadCreationOptions(NoneCreationOptions);
+        var attestationData = DataReader.ReadAttestationData(PackedAttestation);
+        var creationOptions = DataReader.ReadCreationOptions(PackedCreationOptions);
 
         // Act
         var result = await attestation.CompleteRegistration(attestationData, creationOptions, CancellationToken.None);
@@ -103,13 +103,13 @@ internal class NoneAttestationIntegrationTests
     }
 
     [Test]
-    public async Task CompleteRegistration_WhenNoneAttestationUsedTwice_ThenReturnsFailure()
+    public async Task CompleteRegistration_WhenPackedAttestationUsedTwice_ThenReturnsFailure()
     {
         // Arrange
         var attestation = _serviceProvider.GetRequiredService<IAttestation>();
 
-        var attestationData = DataReader.ReadAttestationData(NoneAttestation);
-        var creationOptions = DataReader.ReadCreationOptions(NoneCreationOptions);
+        var attestationData = DataReader.ReadAttestationData(PackedAttestation);
+        var creationOptions = DataReader.ReadCreationOptions(PackedCreationOptions);
 
         await attestation.CompleteRegistration(attestationData, creationOptions, CancellationToken.None);
 
@@ -123,7 +123,7 @@ internal class NoneAttestationIntegrationTests
     }
 
     [Test]
-    public async Task BeginAuthentication_WhenNoneAssertion_ThenReturnsSuccess()
+    public async Task BeginAuthentication_WhenPackedAssertion_ThenReturnsSuccess()
     {
         // Arrange
         var assertion = _serviceProvider.GetRequiredService<IAssertion>();
@@ -154,18 +154,18 @@ internal class NoneAttestationIntegrationTests
     }
 
     [Test]
-    public async Task CompleteAuthentication_WhenNoneAssertion_ThenReturnsSuccess()
+    public async Task CompleteAuthentication_WhenPackedAssertion_ThenReturnsSuccess()
     {
         // Arrange
         var assertion = _serviceProvider.GetRequiredService<IAssertion>();
         var attestation = _serviceProvider.GetRequiredService<IAttestation>();
 
-        var attestationData = DataReader.ReadAttestationData(NoneAttestation);
-        var creationOptions = DataReader.ReadCreationOptions(NoneCreationOptions);
+        var attestationData = DataReader.ReadAttestationData(PackedAttestation);
+        var creationOptions = DataReader.ReadCreationOptions(PackedCreationOptions);
         await attestation.CompleteRegistration(attestationData, creationOptions, CancellationToken.None);
 
-        var assertionData = DataReader.ReadAssertionData(NoneAssertion);
-        var requestOptions = DataReader.ReadRequestOptions(NoneRequestOptions);
+        var assertionData = DataReader.ReadAssertionData(PackedAssertion);
+        var requestOptions = DataReader.ReadRequestOptions(PackedRequestOptions);
 
         // Act
         var result = await assertion.CompleteAuthentication(assertionData, requestOptions, CancellationToken.None);
@@ -177,18 +177,18 @@ internal class NoneAttestationIntegrationTests
     }
 
     [Test]
-    public async Task CompleteAuthentication_WhenNoneAssertionUsedTwice_ThenReturnsFailure()
+    public async Task CompleteAuthentication_WhenPackedAssertionUsedTwice_ThenReturnsFailure()
     {
         // Arrange
         var assertion = _serviceProvider.GetRequiredService<IAssertion>();
         var attestation = _serviceProvider.GetRequiredService<IAttestation>();
 
-        var attestationData = DataReader.ReadAttestationData(NoneAttestation);
-        var creationOptions = DataReader.ReadCreationOptions(NoneCreationOptions);
+        var attestationData = DataReader.ReadAttestationData(PackedAttestation);
+        var creationOptions = DataReader.ReadCreationOptions(PackedCreationOptions);
         await attestation.CompleteRegistration(attestationData, creationOptions, CancellationToken.None);
 
-        var assertionData = DataReader.ReadAssertionData(NoneAssertion);
-        var requestOptions = DataReader.ReadRequestOptions(NoneRequestOptions);
+        var assertionData = DataReader.ReadAssertionData(PackedAssertion);
+        var requestOptions = DataReader.ReadRequestOptions(PackedRequestOptions);
 
         await assertion.CompleteAuthentication(assertionData, requestOptions, CancellationToken.None);
 
@@ -197,7 +197,7 @@ internal class NoneAttestationIntegrationTests
 
         // Assert
         Assert.That(result, Is.Not.Null);
-        Assert.That(result.IsValid, Is.True);
-        Assert.That(result.Message, Is.Null);
+        Assert.That(result.IsValid, Is.False);
+        Assert.That(result.Message, Is.EqualTo("The authenticator's signature counter value is less than or equal to the previously stored count, indicating that the device may have been cloned or duplicated."));
     }
 }
